fix: BaseStyleDefinition abstract class issues
